Fixed all eslint issues in master
