+++ app/src/components/Logo/Logo.jsx
@@ -1,11 +1,10 @@
 import resumakeLogo from '../../assets/images/logo.png'
 import classes from './Logo.module.css'
 
-
 const Logo = () => (
-    <div className={classes.Logo}>
-        <img src={resumakeLogo} alt="ResuMake"/>
-    </div>
+  <div className={classes.Logo}>
+    <img src={resumakeLogo} alt="ResuMake" />
+  </div>
 )
 
 export default Logo
