+++ app/src/components/Navigation/Toolbar/Toolbar.jsx
@@ -1,0 +1,10 @@
+import Logo from '../../Logo/Logo'
+import classes from './Toolbar.module.css'
+
+const Toolbar = () => (
+  <header className={classes.Toolbar}>
+    <Logo />
+  </header>
+)
+
+export default Toolbar;
